* master: (175 commits)
  [tests] fix math tests
  [fixedwing] use min/max_cruise_throttle variables instead of defines
  [conf] update some telemetry files
  [ext] update luftboot
  [ext] update libopencm3
  [mavlink] add sending of origin and waypoints
  [mavlink] rename SETTINGS to SETTINGS_NAMES_SHORT
  [mavlink] send quaternion
  [mavlink] parse PARAM_SET
  [conf] add AUTOPILOT_VERSION to some default telemetry files
  [conf] add python real time plotter to control panel
  [python] messagesapp: growable value column width
  [auto1-flaps] Please remove this as soon a better solution exits. For now it is critical to fly. closing #878
  [conf] vivify fix digital_cam_uart missing settings
  [conf] obc fix nav_airspeed names, and enable AMSL
  fourth release candidate for next stable release v5.4
  [server] close file even if file empty
  [build] improve build version stuff
  Add some more flexibility to conf and control_panel naming
  [paparazzicenter] fix GLib-CRITICAL warning when running clean/build/upload
  ...

Conflicts:
	sw/airborne/boards/ardrone/actuators_at.c
	sw/airborne/firmwares/fixedwing/main_ap.c
	sw/airborne/firmwares/rotorcraft/main.c
	sw/airborne/modules/geo_mag/geo_mag.c
	sw/airborne/modules/ins/ahrs_chimu_spi.c
	sw/airborne/modules/ins/ahrs_chimu_uart.c
	sw/airborne/subsystems/ins/ins_ardrone2.c
	sw/airborne/test/subsystems/test_ahrs.c

+++ sw/airborne/modules/cam_control/point.h
@@ -25,7 +25,7 @@
 #define POINT_H
 
 #if defined(SHOW_CAM_COORDINATES)
-extern unsigned int cam_point_distance_from_home;
+extern uint16_t cam_point_distance_from_home;
 extern float cam_point_lon,  cam_point_lat;
 extern float distance_correction;
 #endif
